+++ DeskHubSharp/Models/License.cs
@@ -1,0 +1,32 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace DeskHubSharp.Models
+{
+    public class License
+    {
+        /// <summary>
+        /// 
+        /// </summary>
+        public string key { get; set; }
+        /// <summary>
+        /// 
+        /// </summary>
+        public string name { get; set; }
+        /// <summary>
+        /// 
+        /// </summary>
+        public string spdx_id { get; set; }
+        /// <summary>
+        /// 
+        /// </summary>
+        public string url { get; set; }
+        /// <summary>
+        /// 
+        /// </summary>
+        public string node_id { get; set; }
+    }
+}
