+++ DeskHubSharp/Models/Owner.cs
@@ -1,0 +1,84 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace DeskHubSharp.Models
+{
+    public class Owner
+    {
+        /// <summary>
+        /// 
+        /// </summary>
+        public string login { get; set; }
+        /// <summary>
+        /// 
+        /// </summary>
+        public int id { get; set; }
+        /// <summary>
+        /// 
+        /// </summary>
+        public string node_id { get; set; }
+        /// <summary>
+        /// 
+        /// </summary>
+        public string avatar_url { get; set; }
+        /// <summary>
+        /// 
+        /// </summary>
+        public string gravatar_id { get; set; }
+        /// <summary>
+        /// 
+        /// </summary>
+        public string url { get; set; }
+        /// <summary>
+        /// 
+        /// </summary>
+        public string html_url { get; set; }
+        /// <summary>
+        /// 
+        /// </summary>
+        public string followers_url { get; set; }
+        /// <summary>
+        /// 
+        /// </summary>
+        public string following_url { get; set; }
+        /// <summary>
+        /// 
+        /// </summary>
+        public string gists_url { get; set; }
+        /// <summary>
+        /// 
+        /// </summary>
+        public string starred_url { get; set; }
+        /// <summary>
+        /// 
+        /// </summary>
+        public string subscriptions_url { get; set; }
+        /// <summary>
+        /// 
+        /// </summary>
+        public string organizations_url { get; set; }
+        /// <summary>
+        /// 
+        /// </summary>
+        public string repos_url { get; set; }
+        /// <summary>
+        /// 
+        /// </summary>
+        public string events_url { get; set; }
+        /// <summary>
+        /// 
+        /// </summary>
+        public string received_events_url { get; set; }
+        /// <summary>
+        /// 
+        /// </summary>
+        public string type { get; set; }
+        /// <summary>
+        /// 
+        /// </summary>
+        public string site_admin { get; set; }
+    }
+}
