+++ DeskHubSharp/SearchWindow.xaml.cs
@@ -33,6 +33,8 @@ namespace DeskHubSharp
         {
             Request request = new Request(txtbox_query.Text);
             request.SearchRequest();
+            //MainWindow main = new MainWindow();
+            //main.ShowDialog();
         }
     }
 }
